Move and refactor filter handlers
