+++ pkg/sqlite/scene_marker_filter.go
@@ -1,0 +1,159 @@
+package sqlite
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/stashapp/stash/pkg/models"
+)
+
+func (qb *SceneMarkerStore) makeFilter(ctx context.Context, sceneMarkerFilter *models.SceneMarkerFilterType) *filterBuilder {
+	query := &filterBuilder{}
+
+	query.handleCriterion(ctx, qb.criterionHandler(sceneMarkerFilter))
+
+	return query
+}
+
+func (qb *SceneMarkerStore) criterionHandler(sceneMarkerFilter *models.SceneMarkerFilterType) criterionHandler {
+	return compoundHandler{
+		qb.tagIDCriterionHandler(sceneMarkerFilter.TagID),
+		qb.tagsCriterionHandler(sceneMarkerFilter.Tags),
+		qb.sceneTagsCriterionHandler(sceneMarkerFilter.SceneTags),
+		qb.performersCriterionHandler(sceneMarkerFilter.Performers),
+		timestampCriterionHandler(sceneMarkerFilter.CreatedAt, "scene_markers.created_at"),
+		timestampCriterionHandler(sceneMarkerFilter.UpdatedAt, "scene_markers.updated_at"),
+		dateCriterionHandler(sceneMarkerFilter.SceneDate, "scenes.date"),
+		timestampCriterionHandler(sceneMarkerFilter.SceneCreatedAt, "scenes.created_at"),
+		timestampCriterionHandler(sceneMarkerFilter.SceneUpdatedAt, "scenes.updated_at"),
+	}
+}
+
+func (qb *SceneMarkerStore) tagIDCriterionHandler(tagID *string) criterionHandlerFunc {
+	return func(ctx context.Context, f *filterBuilder) {
+		if tagID != nil {
+			f.addLeftJoin("scene_markers_tags", "", "scene_markers_tags.scene_marker_id = scene_markers.id")
+
+			f.addWhere("(scene_markers.primary_tag_id = ? OR scene_markers_tags.tag_id = ?)", *tagID, *tagID)
+		}
+	}
+}
+
+func (qb *SceneMarkerStore) tagsCriterionHandler(criterion *models.HierarchicalMultiCriterionInput) criterionHandlerFunc {
+	return func(ctx context.Context, f *filterBuilder) {
+		if criterion != nil {
+			tags := criterion.CombineExcludes()
+
+			if tags.Modifier == models.CriterionModifierIsNull || tags.Modifier == models.CriterionModifierNotNull {
+				var notClause string
+				if tags.Modifier == models.CriterionModifierNotNull {
+					notClause = "NOT"
+				}
+
+				f.addLeftJoin("scene_markers_tags", "", "scene_markers.id = scene_markers_tags.scene_marker_id")
+
+				f.addWhere(fmt.Sprintf("%s scene_markers_tags.tag_id IS NULL", notClause))
+				return
+			}
+
+			if tags.Modifier == models.CriterionModifierEquals && tags.Depth != nil && *tags.Depth != 0 {
+				f.setError(fmt.Errorf("depth is not supported for equals modifier for marker tag filtering"))
+				return
+			}
+
+			if len(tags.Value) == 0 && len(tags.Excludes) == 0 {
+				return
+			}
+
+			if len(tags.Value) > 0 {
+				valuesClause, err := getHierarchicalValues(ctx, qb.tx, tags.Value, tagTable, "tags_relations", "parent_id", "child_id", tags.Depth)
+				if err != nil {
+					f.setError(err)
+					return
+				}
+
+				f.addWith(`marker_tags AS (
+	SELECT mt.scene_marker_id, t.column1 AS root_tag_id FROM scene_markers_tags mt
+	INNER JOIN (` + valuesClause + `) t ON t.column2 = mt.tag_id
+	UNION
+	SELECT m.id, t.column1 FROM scene_markers m
+	INNER JOIN (` + valuesClause + `) t ON t.column2 = m.primary_tag_id
+	)`)
+
+				f.addLeftJoin("marker_tags", "", "marker_tags.scene_marker_id = scene_markers.id")
+
+				switch tags.Modifier {
+				case models.CriterionModifierEquals:
+					// includes only the provided ids
+					f.addWhere("marker_tags.root_tag_id IS NOT NULL")
+					tagsLen := len(tags.Value)
+					f.addHaving(fmt.Sprintf("count(distinct marker_tags.root_tag_id) IS %d", tagsLen))
+					// decrement by one to account for primary tag id
+					f.addWhere("(SELECT COUNT(*) FROM scene_markers_tags s WHERE s.scene_marker_id = scene_markers.id) = ?", tagsLen-1)
+				case models.CriterionModifierNotEquals:
+					f.setError(fmt.Errorf("not equals modifier is not supported for scene marker tags"))
+				default:
+					addHierarchicalConditionClauses(f, tags, "marker_tags", "root_tag_id")
+				}
+			}
+
+			if len(criterion.Excludes) > 0 {
+				valuesClause, err := getHierarchicalValues(ctx, dbWrapper{}, tags.Excludes, tagTable, "tags_relations", "parent_id", "child_id", tags.Depth)
+				if err != nil {
+					f.setError(err)
+					return
+				}
+
+				clause := "scene_markers.id NOT IN (SELECT scene_markers_tags.scene_marker_id FROM scene_markers_tags WHERE scene_markers_tags.tag_id IN (SELECT column2 FROM (%s)))"
+				f.addWhere(fmt.Sprintf(clause, valuesClause))
+
+				f.addWhere(fmt.Sprintf("scene_markers.primary_tag_id NOT IN (SELECT column2 FROM (%s))", valuesClause))
+			}
+		}
+	}
+}
+
+func (qb *SceneMarkerStore) sceneTagsCriterionHandler(tags *models.HierarchicalMultiCriterionInput) criterionHandlerFunc {
+	return func(ctx context.Context, f *filterBuilder) {
+		if tags != nil {
+			f.addLeftJoin("scenes_tags", "", "scene_markers.scene_id = scenes_tags.scene_id")
+
+			h := joinedHierarchicalMultiCriterionHandlerBuilder{
+				tx: qb.tx,
+
+				primaryTable: "scene_markers",
+				primaryKey:   sceneIDColumn,
+				foreignTable: tagTable,
+				foreignFK:    tagIDColumn,
+
+				relationsTable: "tags_relations",
+				joinTable:      "scenes_tags",
+				joinAs:         "marker_scenes_tags",
+				primaryFK:      sceneIDColumn,
+			}
+
+			h.handler(tags).handle(ctx, f)
+		}
+	}
+}
+
+func (qb *SceneMarkerStore) performersCriterionHandler(performers *models.MultiCriterionInput) criterionHandlerFunc {
+	h := joinedMultiCriterionHandlerBuilder{
+		primaryTable: sceneTable,
+		joinTable:    performersScenesTable,
+		joinAs:       "performers_join",
+		primaryFK:    sceneIDColumn,
+		foreignFK:    performerIDColumn,
+
+		addJoinTable: func(f *filterBuilder) {
+			f.addLeftJoin(performersScenesTable, "performers_join", "performers_join.scene_id = scene_markers.scene_id")
+		},
+	}
+
+	handler := h.handler(performers)
+	return func(ctx context.Context, f *filterBuilder) {
+		// Make sure scenes is included, otherwise excludes filter fails
+		f.addLeftJoin(sceneTable, "", "scenes.id = scene_markers.scene_id")
+		handler(ctx, f)
+	}
+}
