+++ pkg/sqlite/studio.go
@@ -452,83 +452,6 @@ func (qb *StudioStore) QueryForAutoTag(ctx context.Context, words []string) ([]*
 	return ret, nil
 }
 
-func (qb *StudioStore) validateFilter(filter *models.StudioFilterType) error {
-	const and = "AND"
-	const or = "OR"
-	const not = "NOT"
-
-	if filter.And != nil {
-		if filter.Or != nil {
-			return illegalFilterCombination(and, or)
-		}
-		if filter.Not != nil {
-			return illegalFilterCombination(and, not)
-		}
-
-		return qb.validateFilter(filter.And)
-	}
-
-	if filter.Or != nil {
-		if filter.Not != nil {
-			return illegalFilterCombination(or, not)
-		}
-
-		return qb.validateFilter(filter.Or)
-	}
-
-	if filter.Not != nil {
-		return qb.validateFilter(filter.Not)
-	}
-
-	return nil
-}
-
-func (qb *StudioStore) makeFilter(ctx context.Context, studioFilter *models.StudioFilterType) *filterBuilder {
-	query := &filterBuilder{}
-
-	if studioFilter.And != nil {
-		query.and(qb.makeFilter(ctx, studioFilter.And))
-	}
-	if studioFilter.Or != nil {
-		query.or(qb.makeFilter(ctx, studioFilter.Or))
-	}
-	if studioFilter.Not != nil {
-		query.not(qb.makeFilter(ctx, studioFilter.Not))
-	}
-
-	query.handleCriterion(ctx, stringCriterionHandler(studioFilter.Name, studioTable+".name"))
-	query.handleCriterion(ctx, stringCriterionHandler(studioFilter.Details, studioTable+".details"))
-	query.handleCriterion(ctx, stringCriterionHandler(studioFilter.URL, studioTable+".url"))
-	query.handleCriterion(ctx, intCriterionHandler(studioFilter.Rating100, studioTable+".rating", nil))
-	query.handleCriterion(ctx, boolCriterionHandler(studioFilter.Favorite, studioTable+".favorite", nil))
-	query.handleCriterion(ctx, boolCriterionHandler(studioFilter.IgnoreAutoTag, studioTable+".ignore_auto_tag", nil))
-
-	query.handleCriterion(ctx, criterionHandlerFunc(func(ctx context.Context, f *filterBuilder) {
-		if studioFilter.StashID != nil {
-			qb.stashIDRepository().join(f, "studio_stash_ids", "studios.id")
-			stringCriterionHandler(studioFilter.StashID, "studio_stash_ids.stash_id")(ctx, f)
-		}
-	}))
-	query.handleCriterion(ctx, &stashIDCriterionHandler{
-		c:                 studioFilter.StashIDEndpoint,
-		stashIDRepository: qb.stashIDRepository(),
-		stashIDTableAs:    "studio_stash_ids",
-		parentIDCol:       "studios.id",
-	})
-
-	query.handleCriterion(ctx, studioIsMissingCriterionHandler(qb, studioFilter.IsMissing))
-	query.handleCriterion(ctx, studioSceneCountCriterionHandler(qb, studioFilter.SceneCount))
-	query.handleCriterion(ctx, studioImageCountCriterionHandler(qb, studioFilter.ImageCount))
-	query.handleCriterion(ctx, studioGalleryCountCriterionHandler(qb, studioFilter.GalleryCount))
-	query.handleCriterion(ctx, studioParentCriterionHandler(qb, studioFilter.Parents))
-	query.handleCriterion(ctx, studioAliasCriterionHandler(qb, studioFilter.Aliases))
-	query.handleCriterion(ctx, studioChildCountCriterionHandler(qb, studioFilter.ChildCount))
-	query.handleCriterion(ctx, timestampCriterionHandler(studioFilter.CreatedAt, studioTable+".created_at"))
-	query.handleCriterion(ctx, timestampCriterionHandler(studioFilter.UpdatedAt, studioTable+".updated_at"))
-
-	return query
-}
-
 func (qb *StudioStore) makeQuery(ctx context.Context, studioFilter *models.StudioFilterType, findFilter *models.FindFilterType) (*queryBuilder, error) {
 	if studioFilter == nil {
 		studioFilter = &models.StudioFilterType{}
@@ -582,93 +505,6 @@ func (qb *StudioStore) Query(ctx context.Context, studioFilter *models.StudioFil
 	}
 
 	return studios, countResult, nil
-}
-
-func studioIsMissingCriterionHandler(qb *StudioStore, isMissing *string) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if isMissing != nil && *isMissing != "" {
-			switch *isMissing {
-			case "image":
-				f.addWhere("studios.image_blob IS NULL")
-			case "stash_id":
-				qb.stashIDRepository().join(f, "studio_stash_ids", "studios.id")
-				f.addWhere("studio_stash_ids.studio_id IS NULL")
-			default:
-				f.addWhere("(studios." + *isMissing + " IS NULL OR TRIM(studios." + *isMissing + ") = '')")
-			}
-		}
-	}
-}
-
-func studioSceneCountCriterionHandler(qb *StudioStore, sceneCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if sceneCount != nil {
-			f.addLeftJoin("scenes", "", "scenes.studio_id = studios.id")
-			clause, args := getIntCriterionWhereClause("count(distinct scenes.id)", *sceneCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func studioImageCountCriterionHandler(qb *StudioStore, imageCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if imageCount != nil {
-			f.addLeftJoin("images", "", "images.studio_id = studios.id")
-			clause, args := getIntCriterionWhereClause("count(distinct images.id)", *imageCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func studioGalleryCountCriterionHandler(qb *StudioStore, galleryCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if galleryCount != nil {
-			f.addLeftJoin("galleries", "", "galleries.studio_id = studios.id")
-			clause, args := getIntCriterionWhereClause("count(distinct galleries.id)", *galleryCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func studioParentCriterionHandler(qb *StudioStore, parents *models.MultiCriterionInput) criterionHandlerFunc {
-	addJoinsFunc := func(f *filterBuilder) {
-		f.addLeftJoin("studios", "parent_studio", "parent_studio.id = studios.parent_id")
-	}
-	h := multiCriterionHandlerBuilder{
-		primaryTable: studioTable,
-		foreignTable: "parent_studio",
-		joinTable:    "",
-		primaryFK:    studioIDColumn,
-		foreignFK:    "parent_id",
-		addJoinsFunc: addJoinsFunc,
-	}
-	return h.handler(parents)
-}
-
-func studioAliasCriterionHandler(qb *StudioStore, alias *models.StringCriterionInput) criterionHandlerFunc {
-	h := stringListCriterionHandlerBuilder{
-		joinTable:    studioAliasesTable,
-		stringColumn: studioAliasColumn,
-		addJoinTable: func(f *filterBuilder) {
-			studiosAliasesTableMgr.join(f, "", "studios.id")
-		},
-	}
-
-	return h.handler(alias)
-}
-
-func studioChildCountCriterionHandler(qb *StudioStore, childCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if childCount != nil {
-			f.addLeftJoin("studios", "children_count", "children_count.parent_id = studios.id")
-			clause, args := getIntCriterionWhereClause("count(distinct children_count.id)", *childCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
 }
 
 var studioSortOptions = sortOptions{
