+++ pkg/sqlite/tag.go
@@ -455,73 +455,6 @@ func (qb *TagStore) QueryForAutoTag(ctx context.Context, words []string) ([]*mod
 	return qb.queryTags(ctx, query+" WHERE "+where, args)
 }
 
-func (qb *TagStore) validateFilter(tagFilter *models.TagFilterType) error {
-	const and = "AND"
-	const or = "OR"
-	const not = "NOT"
-
-	if tagFilter.And != nil {
-		if tagFilter.Or != nil {
-			return illegalFilterCombination(and, or)
-		}
-		if tagFilter.Not != nil {
-			return illegalFilterCombination(and, not)
-		}
-
-		return qb.validateFilter(tagFilter.And)
-	}
-
-	if tagFilter.Or != nil {
-		if tagFilter.Not != nil {
-			return illegalFilterCombination(or, not)
-		}
-
-		return qb.validateFilter(tagFilter.Or)
-	}
-
-	if tagFilter.Not != nil {
-		return qb.validateFilter(tagFilter.Not)
-	}
-
-	return nil
-}
-
-func (qb *TagStore) makeFilter(ctx context.Context, tagFilter *models.TagFilterType) *filterBuilder {
-	query := &filterBuilder{}
-
-	if tagFilter.And != nil {
-		query.and(qb.makeFilter(ctx, tagFilter.And))
-	}
-	if tagFilter.Or != nil {
-		query.or(qb.makeFilter(ctx, tagFilter.Or))
-	}
-	if tagFilter.Not != nil {
-		query.not(qb.makeFilter(ctx, tagFilter.Not))
-	}
-
-	query.handleCriterion(ctx, stringCriterionHandler(tagFilter.Name, tagTable+".name"))
-	query.handleCriterion(ctx, tagAliasCriterionHandler(qb, tagFilter.Aliases))
-
-	query.handleCriterion(ctx, boolCriterionHandler(tagFilter.Favorite, tagTable+".favorite", nil))
-	query.handleCriterion(ctx, stringCriterionHandler(tagFilter.Description, tagTable+".description"))
-	query.handleCriterion(ctx, boolCriterionHandler(tagFilter.IgnoreAutoTag, tagTable+".ignore_auto_tag", nil))
-
-	query.handleCriterion(ctx, tagIsMissingCriterionHandler(qb, tagFilter.IsMissing))
-	query.handleCriterion(ctx, tagSceneCountCriterionHandler(qb, tagFilter.SceneCount))
-	query.handleCriterion(ctx, tagImageCountCriterionHandler(qb, tagFilter.ImageCount))
-	query.handleCriterion(ctx, tagGalleryCountCriterionHandler(qb, tagFilter.GalleryCount))
-	query.handleCriterion(ctx, tagPerformerCountCriterionHandler(qb, tagFilter.PerformerCount))
-	query.handleCriterion(ctx, tagMarkerCountCriterionHandler(qb, tagFilter.MarkerCount))
-	query.handleCriterion(ctx, tagParentsCriterionHandler(qb, tagFilter.Parents))
-	query.handleCriterion(ctx, tagChildrenCriterionHandler(qb, tagFilter.Children))
-	query.handleCriterion(ctx, tagParentCountCriterionHandler(qb, tagFilter.ParentCount))
-	query.handleCriterion(ctx, tagChildCountCriterionHandler(qb, tagFilter.ChildCount))
-	query.handleCriterion(ctx, timestampCriterionHandler(tagFilter.CreatedAt, "tags.created_at"))
-	query.handleCriterion(ctx, timestampCriterionHandler(tagFilter.UpdatedAt, "tags.updated_at"))
-
-	return query
-}
-
 func (qb *TagStore) Query(ctx context.Context, tagFilter *models.TagFilterType, findFilter *models.FindFilterType) ([]*models.Tag, int, error) {
 	if tagFilter == nil {
 		tagFilter = &models.TagFilterType{}
@@ -565,297 +498,6 @@ func (qb *TagStore) Query(ctx context.Context, tagFilter *models.TagFilterType, 
 	}
 
 	return tags, countResult, nil
-}
-
-func tagAliasCriterionHandler(qb *TagStore, alias *models.StringCriterionInput) criterionHandlerFunc {
-	h := stringListCriterionHandlerBuilder{
-		joinTable:    tagAliasesTable,
-		stringColumn: tagAliasColumn,
-		addJoinTable: func(f *filterBuilder) {
-			qb.aliasRepository().join(f, "", "tags.id")
-		},
-	}
-
-	return h.handler(alias)
-}
-
-func tagIsMissingCriterionHandler(qb *TagStore, isMissing *string) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if isMissing != nil && *isMissing != "" {
-			switch *isMissing {
-			case "image":
-				f.addWhere("tags.image_blob IS NULL")
-			default:
-				f.addWhere("(tags." + *isMissing + " IS NULL OR TRIM(tags." + *isMissing + ") = '')")
-			}
-		}
-	}
-}
-
-func tagSceneCountCriterionHandler(qb *TagStore, sceneCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if sceneCount != nil {
-			f.addLeftJoin("scenes_tags", "", "scenes_tags.tag_id = tags.id")
-			clause, args := getIntCriterionWhereClause("count(distinct scenes_tags.scene_id)", *sceneCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func tagImageCountCriterionHandler(qb *TagStore, imageCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if imageCount != nil {
-			f.addLeftJoin("images_tags", "", "images_tags.tag_id = tags.id")
-			clause, args := getIntCriterionWhereClause("count(distinct images_tags.image_id)", *imageCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func tagGalleryCountCriterionHandler(qb *TagStore, galleryCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if galleryCount != nil {
-			f.addLeftJoin("galleries_tags", "", "galleries_tags.tag_id = tags.id")
-			clause, args := getIntCriterionWhereClause("count(distinct galleries_tags.gallery_id)", *galleryCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func tagPerformerCountCriterionHandler(qb *TagStore, performerCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if performerCount != nil {
-			f.addLeftJoin("performers_tags", "", "performers_tags.tag_id = tags.id")
-			clause, args := getIntCriterionWhereClause("count(distinct performers_tags.performer_id)", *performerCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func tagMarkerCountCriterionHandler(qb *TagStore, markerCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if markerCount != nil {
-			f.addLeftJoin("scene_markers_tags", "", "scene_markers_tags.tag_id = tags.id")
-			f.addLeftJoin("scene_markers", "", "scene_markers_tags.scene_marker_id = scene_markers.id OR scene_markers.primary_tag_id = tags.id")
-			clause, args := getIntCriterionWhereClause("count(distinct scene_markers.id)", *markerCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func tagParentsCriterionHandler(qb *TagStore, criterion *models.HierarchicalMultiCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if criterion != nil {
-			tags := criterion.CombineExcludes()
-
-			// validate the modifier
-			switch tags.Modifier {
-			case models.CriterionModifierIncludesAll, models.CriterionModifierIncludes, models.CriterionModifierExcludes, models.CriterionModifierIsNull, models.CriterionModifierNotNull:
-				// valid
-			default:
-				f.setError(fmt.Errorf("invalid modifier %s for tag parent/children", criterion.Modifier))
-			}
-
-			if tags.Modifier == models.CriterionModifierIsNull || tags.Modifier == models.CriterionModifierNotNull {
-				var notClause string
-				if tags.Modifier == models.CriterionModifierNotNull {
-					notClause = "NOT"
-				}
-
-				f.addLeftJoin("tags_relations", "parent_relations", "tags.id = parent_relations.child_id")
-
-				f.addWhere(fmt.Sprintf("parent_relations.parent_id IS %s NULL", notClause))
-				return
-			}
-
-			if len(tags.Value) == 0 && len(tags.Excludes) == 0 {
-				return
-			}
-
-			if len(tags.Value) > 0 {
-				var args []interface{}
-				for _, val := range tags.Value {
-					args = append(args, val)
-				}
-
-				depthVal := 0
-				if tags.Depth != nil {
-					depthVal = *tags.Depth
-				}
-
-				var depthCondition string
-				if depthVal != -1 {
-					depthCondition = fmt.Sprintf("WHERE depth < %d", depthVal)
-				}
-
-				query := `parents AS (
-		SELECT parent_id AS root_id, child_id AS item_id, 0 AS depth FROM tags_relations WHERE parent_id IN` + getInBinding(len(tags.Value)) + `
-		UNION
-		SELECT root_id, child_id, depth + 1 FROM tags_relations INNER JOIN parents ON item_id = parent_id ` + depthCondition + `
-	)`
-
-				f.addRecursiveWith(query, args...)
-
-				f.addLeftJoin("parents", "", "parents.item_id = tags.id")
-
-				addHierarchicalConditionClauses(f, tags, "parents", "root_id")
-			}
-
-			if len(tags.Excludes) > 0 {
-				var args []interface{}
-				for _, val := range tags.Excludes {
-					args = append(args, val)
-				}
-
-				depthVal := 0
-				if tags.Depth != nil {
-					depthVal = *tags.Depth
-				}
-
-				var depthCondition string
-				if depthVal != -1 {
-					depthCondition = fmt.Sprintf("WHERE depth < %d", depthVal)
-				}
-
-				query := `parents2 AS (
-		SELECT parent_id AS root_id, child_id AS item_id, 0 AS depth FROM tags_relations WHERE parent_id IN` + getInBinding(len(tags.Excludes)) + `
-		UNION
-		SELECT root_id, child_id, depth + 1 FROM tags_relations INNER JOIN parents2 ON item_id = parent_id ` + depthCondition + `
-	)`
-
-				f.addRecursiveWith(query, args...)
-
-				f.addLeftJoin("parents2", "", "parents2.item_id = tags.id")
-
-				addHierarchicalConditionClauses(f, models.HierarchicalMultiCriterionInput{
-					Value:    tags.Excludes,
-					Depth:    tags.Depth,
-					Modifier: models.CriterionModifierExcludes,
-				}, "parents2", "root_id")
-			}
-		}
-	}
-}
-
-func tagChildrenCriterionHandler(qb *TagStore, criterion *models.HierarchicalMultiCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if criterion != nil {
-			tags := criterion.CombineExcludes()
-
-			// validate the modifier
-			switch tags.Modifier {
-			case models.CriterionModifierIncludesAll, models.CriterionModifierIncludes, models.CriterionModifierExcludes, models.CriterionModifierIsNull, models.CriterionModifierNotNull:
-				// valid
-			default:
-				f.setError(fmt.Errorf("invalid modifier %s for tag parent/children", criterion.Modifier))
-			}
-
-			if tags.Modifier == models.CriterionModifierIsNull || tags.Modifier == models.CriterionModifierNotNull {
-				var notClause string
-				if tags.Modifier == models.CriterionModifierNotNull {
-					notClause = "NOT"
-				}
-
-				f.addLeftJoin("tags_relations", "child_relations", "tags.id = child_relations.parent_id")
-
-				f.addWhere(fmt.Sprintf("child_relations.child_id IS %s NULL", notClause))
-				return
-			}
-
-			if len(tags.Value) == 0 && len(tags.Excludes) == 0 {
-				return
-			}
-
-			if len(tags.Value) > 0 {
-				var args []interface{}
-				for _, val := range tags.Value {
-					args = append(args, val)
-				}
-
-				depthVal := 0
-				if tags.Depth != nil {
-					depthVal = *tags.Depth
-				}
-
-				var depthCondition string
-				if depthVal != -1 {
-					depthCondition = fmt.Sprintf("WHERE depth < %d", depthVal)
-				}
-
-				query := `children AS (
-		SELECT child_id AS root_id, parent_id AS item_id, 0 AS depth FROM tags_relations WHERE child_id IN` + getInBinding(len(tags.Value)) + `
-		UNION
-		SELECT root_id, parent_id, depth + 1 FROM tags_relations INNER JOIN children ON item_id = child_id ` + depthCondition + `
-	)`
-
-				f.addRecursiveWith(query, args...)
-
-				f.addLeftJoin("children", "", "children.item_id = tags.id")
-
-				addHierarchicalConditionClauses(f, tags, "children", "root_id")
-			}
-
-			if len(tags.Excludes) > 0 {
-				var args []interface{}
-				for _, val := range tags.Excludes {
-					args = append(args, val)
-				}
-
-				depthVal := 0
-				if tags.Depth != nil {
-					depthVal = *tags.Depth
-				}
-
-				var depthCondition string
-				if depthVal != -1 {
-					depthCondition = fmt.Sprintf("WHERE depth < %d", depthVal)
-				}
-
-				query := `children2 AS (
-		SELECT child_id AS root_id, parent_id AS item_id, 0 AS depth FROM tags_relations WHERE child_id IN` + getInBinding(len(tags.Excludes)) + `
-		UNION
-		SELECT root_id, parent_id, depth + 1 FROM tags_relations INNER JOIN children2 ON item_id = child_id ` + depthCondition + `
-	)`
-
-				f.addRecursiveWith(query, args...)
-
-				f.addLeftJoin("children2", "", "children2.item_id = tags.id")
-
-				addHierarchicalConditionClauses(f, models.HierarchicalMultiCriterionInput{
-					Value:    tags.Excludes,
-					Depth:    tags.Depth,
-					Modifier: models.CriterionModifierExcludes,
-				}, "children2", "root_id")
-			}
-		}
-	}
-}
-
-func tagParentCountCriterionHandler(qb *TagStore, parentCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if parentCount != nil {
-			f.addLeftJoin("tags_relations", "parents_count", "parents_count.child_id = tags.id")
-			clause, args := getIntCriterionWhereClause("count(distinct parents_count.parent_id)", *parentCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
-}
-
-func tagChildCountCriterionHandler(qb *TagStore, childCount *models.IntCriterionInput) criterionHandlerFunc {
-	return func(ctx context.Context, f *filterBuilder) {
-		if childCount != nil {
-			f.addLeftJoin("tags_relations", "children_count", "children_count.parent_id = tags.id")
-			clause, args := getIntCriterionWhereClause("count(distinct children_count.child_id)", *childCount)
-
-			f.addHaving(clause, args...)
-		}
-	}
 }
 
 var tagSortOptions = sortOptions{
